+++ CHANGES.rst
@@ -4,9 +4,11 @@ OpenREM version history
 
 0.8.2 (2018-xx-xx)
 ------------------
+* `#689`_  Interface: fixed URL lookup error for DICOM summary page
 * `#685`_  Charts: fixed link code that would otherwise cause DLP per acqusition protocol chart histogram links to fail
 * `#683`_  Installation: added VIRTUAL_DIRECTORY to the settings file to avoid updating local_settings file on upgrade
 * `#682`_  Charts: fixed problem where links from histogram bars didn't filter correctly when case-insensitive categories selected
+* `#679`_  Interface: added ability to filter CT studies on acquisition type
 * `#677`_  Interface: added additional filter materials to convert to abbreviations
 * `#675`_  Exports: improved resilience when export includes malformed studies
 * `#674`_  Documentation: amended zip command in example Orthanc configuration to work with Linux and Windows